fix: move schemas to the separate module
